Revert TypeScripting dropzone-widget (infinite render issue)
